requirements.txt
anyascii==0.1.7
asgiref==3.3.4
beautifulsoup4==4.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.8
django-allauth==0.43.0
django-crispy-forms==1.9.2
django-environ==0.4.5
django-filter==2.4.0
django-modelcluster==5.1
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.12.1
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
geographiclib==1.50
geopy==2.0.0
googlemaps==4.4.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
jdcal==1.4.1
l18n==2020.6.1
oauthlib==3.1.0
openpyxl==3.0.5
Pillow==7.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==2.0.0
Unidecode==1.1.1
urllib3==1.25.11
wagtail==2.12.3
webencodings==0.5.1
Willow==1.4
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
